Backend/requirements.txt
fastapi
uvicorn
python-multipart
moviepy
opencv-python
openai
openai-whisper
